Flush redis cache on save, if redis is installed

+++ src/php/functions.php
@@ -1,11 +1,15 @@
 <?php
 
-/* function flush_redis() {
+function flush_redis() {
+    if (!class_exists('Redis')) {
+        return;
+    }
+
     $redis = new Redis();
     $redis->connect('127.0.0.1');
     $redis->flushAll();
 }
-add_action('save_post', 'flush_redis'); */
+add_action('save_post', 'flush_redis');
 
 $dev_fake_ie = false;
